+++ src/C#/ChickenSharp/ChickenInterfaces/IInstructionSet.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Esoterics.Interfaces
+namespace Esoterics.ChickenInterfaces
 {
     public interface IInstructionSet
     {
