+++ src/C#/ChickenSharp/ChickenInterfaces/IStack.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Esoterics.Interfaces
+namespace Esoterics.ChickenInterfaces
 {
     public interface IStack
     {
