+++ src/C#/ChickenSharp/ChickenInterfaces/IVM.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Esoterics.Interfaces
+namespace Esoterics.ChickenInterfaces
 {
     public interface IVM
     {
